[feat] Tests folder moved outside of pypipe folder
